refactor(repository): extract code into abstract class
Most of the code in `PhpEntityRepository` is relevant
for other repositories based on an entity provider
too. Hence, the re-usable logic was moved into a
newly created abstract class.

Refs: #168
